+++ test/Compze.Tests.Integration/InProcess/Given_an_endpoint_declaring_only_in_process_tessaging_and_typermedia.cs
@@ -62,8 +62,8 @@ public class Given_an_endpoint_declaring_only_in_process_tessaging_and_typermedi
 
    [PCT] public void a_tevent_published_in_process_reaches_the_endpoints_subscriber()
    {
-      _endpoint.ServiceLocator.Resolve<IScopeFactory>().ExecuteInIsolatedScope(scope =>
-         scope.Resolve<IUnitOfWorkTeventPublisher>().Publish(new MySpecialGreetingRequestedTevent()));
+      _endpoint.ServiceLocator.Resolve<IScopeFactory>().ExecuteUnitOfWork(unitOfWork =>
+         unitOfWork.Resolve<IUnitOfWorkTeventPublisher>().Publish(new MySpecialGreetingRequestedTevent()));
       _receivedTevents.Must().HaveCount(1);
    }
 }